- Modified views.py to handle authentication properly. - Refactored TaskViewSet to automatically assign new tasks to the currently authenticated user.

+++ tasks/views.py
@@ -1,43 +1,36 @@
-from rest_framework import viewsets, status
+from rest_framework import viewsets, status, generics
 from rest_framework.response import Response
 from django.contrib.auth.models import User
 from .models import Task
-from .serializers import TaskSerializer, UserSerializer
+from .serializers import TaskSerializer, UserRegistrationSerializer
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.exceptions import PermissionDenied
+
+class UserRegistrationView(generics.CreateAPIView):
+    #API endpoint for user registration
+    serializer_class = UserRegistrationSerializer
+    permission_classes = [] 
 
 class TaskViewSet(viewsets.ModelViewSet):
     queryset = Task.objects.all()
     serializer_class = TaskSerializer
+    permission_classes = [IsAuthenticated]
     
-    def create(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        if serializer.is_valid():
+    def get_queryset(self):
+        user = self.request.user
+        return Task.objects.filter(assigned_to=user)
+        
+    def perform_create(self, serializer):
+        serializer.save(assigned_to=self.request.user)
+    
+    def perform_update(self, serializer):
+        if serializer.instance.assigned_to == self.request.user:
             serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def list(self, request, *args, **kwargs):
-        queryset = self.get_queryset()
-        serializer = self.get_serializer(queryset, many=True)
-        return Response(serializer.data)
-    
-    def retrieve(self, request, *args, **kwargs):
-        instance = self.get_object()
-        serializer = self.get_serializer(instance)
-        return Response(serializer.data)
-    
-    def update(self, request, *args, **kwargs):
-        instance = self.get_object()
-        serializer = self.get_serializer(instance, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def destroy(self, request, *args, **kwargs):
-        instance = self.get_object()
-        instance.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            raise PermissionDenied("You do not have permission to edit this task.")
 
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
+    def perform_destroy(self, instance):
+        if instance.assigned_to == self.request.user:
+            instance.delete()
+        else:
+            raise PermissionDenied("You do not have permission to delete this task.")
